- Save button working - Organization of the code in respective folder - Card for save recipe initialized
