+++ app/src/main/java/com/example/whattoeat/data/entities/SaveRecipeEntity.kt
@@ -1,6 +1,5 @@
 package com.example.whattoeat.data.entities
 
-import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.Index
 import androidx.room.PrimaryKey
@@ -8,7 +7,7 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 @Entity(tableName = "save_recipe_table", indices = [Index(value = ["idRecipe"], unique = true)])
-data class RecipeSaveEntity(
+data class SaveRecipeEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Int = 0,
     val idRecipe: Int = 0,
