+++ app/src/main/java/com/example/whattoeat/ui/theme/screens/saveRecipe/SaveRecipeScreen.kt
@@ -1,17 +1,20 @@
 package com.example.whattoeat.ui.theme.screens.saveRecipe
 
-import androidx.activity.viewModels
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.rememberNavController
-import com.example.whattoeat.data.entities.RecipeSaveEntity
+import com.example.whattoeat.data.entities.SaveRecipeEntity
+import com.example.whattoeat.ui.theme.screens.randomRecipe.components.IngredientCard
+import com.example.whattoeat.ui.theme.screens.saveRecipe.components.SaveRecipeCard
 
 
 @Composable
@@ -29,8 +32,18 @@ fun SaveRecipeScreen(
 
 
 @Composable
-fun test(saveRecipes: List<RecipeSaveEntity>) {
+fun test(saveRecipes: List<SaveRecipeEntity>) {
     Surface(modifier = Modifier.fillMaxSize()) {
-        Text(text = saveRecipes[0].title.toString())
+        LazyColumn(
+            modifier = Modifier
+                .fillMaxWidth(0.95f)
+              ,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            items(saveRecipes) { recipe ->
+                SaveRecipeCard(saveRecipe = recipe)
+            }
+        }
+
     }
 }
