+++ app/src/main/java/com/example/whattoeat/ui/theme/screens/saveRecipe/SaveRecipeViewModel.kt
@@ -1,15 +1,13 @@
 package com.example.whattoeat.ui.theme.screens.saveRecipe
 
-import android.app.Application
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.whattoeat.data.entities.RecipeSaveEntity
+import com.example.whattoeat.data.entities.SaveRecipeEntity
 import com.example.whattoeat.data.repositories.SaveRecipeRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -17,7 +15,7 @@ import javax.inject.Inject
 class SaveRecipeViewModel @Inject constructor(private val _saveRecipeRepository:SaveRecipeRepository): ViewModel() {
 
 
-    var saveRecipe by mutableStateOf(RecipeSaveEntity())
+    var saveRecipe by mutableStateOf(SaveRecipeEntity())
         private set
 
     val getAllSaveRecipe = _saveRecipeRepository.getAllSaveRecipe()
